# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  request.HostToolConfig:
    properties:
      content:
        type: string
      operate:
        enum:
        - get
        - set
        type: string
      type:
        enum:
        - supervisord
        type: string
    required:
    - type
    type: object
  request.HostToolCreate:
    properties:
      configPath:
        type: string
      serviceName:
        type: string
      type:
        type: string
    required:
    - type
    type: object
  request.HostToolLogReq:
    properties:
      type:
        enum:
        - supervisord
        type: string
    required:
    - type
    type: object
  request.HostToolReq:
    properties:
      operate:
        enum:
        - status
        - restart
        - start
        - stop
        type: string
      type:
        enum:
        - supervisord
        type: string
    required:
    - type
    type: object
...
  request.SupervisorProcessConfig:
    properties:
      command:
        type: string
      dir:
        type: string
      name:
        type: string
      numprocs:
        type: string
      operate:
        type: string
      user:
        type: string
    type: object
  request.SupervisorProcessFileReq:
    properties:
      content:
        type: string
      file:
        enum:
        - out.log
        - err.log
        - config
        type: string
      name:
        type: string
      operate:
        enum:
        - get
        - clear
        - update
        type: string
    required:
    - file
    - name
    - operate
    type: object
...
host: localhost
info:
  contact: {}
  description: 开源Linux面板
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 1Panel
  version: "1.0"
paths:
...
  /host/tool:
    post:
      consumes:
      - application/json
      description: 获取主机工具状态
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.HostToolReq'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get tool
      tags:
      - Host tool
  /host/tool/config:
    post:
      consumes:
      - application/json
      description: 操作主机工具配置文件
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.HostToolConfig'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get tool config
      tags:
      - Host tool
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - operate
        formatEN: '[operate] tool config'
        formatZH: '[operate] 主机工具配置文件 '
        paramKeys: []
  /host/tool/create:
    post:
      consumes:
      - application/json
      description: 创建主机工具配置
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.HostToolCreate'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create Host tool Config
      tags:
      - Host tool
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - type
        formatEN: create [type] config
        formatZH: 创建 [type] 配置
        paramKeys: []
  /host/tool/log:
    post:
      consumes:
      - application/json
      description: 获取主机工具日志
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.HostToolLogReq'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get tool
      tags:
      - Host tool
  /host/tool/operate:
    post:
      consumes:
      - application/json
      description: 操作主机工具
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.HostToolReq'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Operate tool
      tags:
      - Host tool
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - operate
        - type
        formatEN: '[operate] [type]'
        formatZH: '[operate] [type] '
        paramKeys: []
  /host/tool/supervisor/process:
    get:
      consumes:
      - application/json
      description: 获取 Supervisor 进程配置
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get Supervisor process config
      tags:
      - Host tool
    post:
      consumes:
      - application/json
      description: 操作守护进程
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.SupervisorProcessConfig'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create Supervisor process
      tags:
      - Host tool
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - operate
        formatEN: '[operate] process'
        formatZH: '[operate] 守护进程 '
        paramKeys: []
  /host/tool/supervisor/process/file:
    post:
      consumes:
      - application/json
      description: 操作 Supervisor 进程文件
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.SupervisorProcessFileReq'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get Supervisor process config
      tags:
      - Host tool
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - operate
        formatEN: '[operate] Supervisor Process Config file'
        formatZH: '[operate] Supervisor 进程文件 '
        paramKeys: []
...
swagger: "2.0"